updated upload captcha and implemented progress info

+++ src/main/java/com/springreact/template/service/StorageService.java
@@ -4,12 +4,16 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
+import com.amazonaws.services.s3.transfer.TransferManager;
+import com.amazonaws.services.s3.transfer.TransferManagerBuilder;
 import com.amazonaws.util.IOUtils;
+import com.springreact.template.controller.HomeController;
 import com.springreact.template.db.Upload;
 import com.springreact.template.db.UploadRepository;
 import com.springreact.template.db.User;
 import com.springreact.template.db.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.tika.Tika;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -19,6 +23,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Date;
+import java.util.concurrent.Executors;
 
 @Service
 @Slf4j
@@ -37,17 +42,37 @@ public class StorageService {
     private UploadRepository uploadRepository;
 
     public String uploadFile(MultipartFile file, User user) {
-        // TODO: Validate File ending
+
         File fileObject = convertMultiPartFileToFile(file);
+        if(fileObject == null)
+            return "";
         String fileName = System.currentTimeMillis()+"_"+file.getOriginalFilename();
 
-        // TODO: Try Catch Block with putObject
-        s3Client.putObject(new PutObjectRequest(bucketName, fileName, fileObject));
-        fileObject.delete();
-        int uploadNumber = (int)Math.floor(Math.random()*(99999-11111+1)+11111);
-        String tempName = "My Upload #" + uploadNumber;
-        Upload up = uploadRepository.save(new Upload(tempName, fileName, new Date(), user, null));
-        return "{ \"url\": \"/api/uploads/" + up.getId() + "\" }";
+        int maxUploadThreads = 5;
+        TransferManager tm = TransferManagerBuilder
+                .standard()
+                .withS3Client(s3Client)
+                .withMultipartUploadThreshold((long) (5 * 1024 * 1024))
+                .withExecutorFactory(() -> Executors.newFixedThreadPool(maxUploadThreads))
+                .build();
+
+        PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, fileName, fileObject);
+
+        try {
+            tm.upload(putObjectRequest).waitForCompletion();
+            fileObject.delete(); // delete temporary file
+            HomeController.validated = false;
+            // save to database
+            int uploadNumber = (int)Math.floor(Math.random()*(99999-11111+1)+11111);
+            String tempName = "My Upload #" + uploadNumber;
+            Upload up = uploadRepository.save(new Upload(tempName, fileName, new Date(), user, null));
+            return "{ \"url\": \"/api/uploads/" + up.getId() + "\" }";
+        }
+        catch (InterruptedException e) {
+            HomeController.validated = false;
+            e.printStackTrace();
+            return "{ \"message\": \"" + e.getMessage() + "\" }";
+        }
     }
 
     public byte[] downloadFile(String fileName) {
@@ -67,12 +92,31 @@ public class StorageService {
     }
 
     private File convertMultiPartFileToFile(MultipartFile file) {
+
+        if(file.isEmpty())
+            return null;
+
+        try {
+            Tika tika = new Tika();
+            // File Types see: https://tika.apache.org/1.26/formats.html
+            if(!tika.detect(file.getBytes()).equals("application/pdf")) {
+                return null;
+            }
+        } catch(IOException e) {
+            System.out.print("### Could not getBytes for Tika validation: ");
+            System.out.println(e.getMessage());
+            return null;
+        }
+
         File convertedFile = new File(file.getOriginalFilename());
         try (FileOutputStream fos = new FileOutputStream(convertedFile)) {
             fos.write(file.getBytes());
         } catch (IOException e) {
-            log.error("Error converting multipartFile to file", e);
+            System.out.print("### Could not Parse Multipart to File: ");
+            System.out.println(e.getMessage());
+            return null;
         }
+
         return convertedFile;
     }
